Notes/10_Pro_Django/02_Logging/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
